car/pyproject.toml
[project]
name = "car"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "numpy>=2.3.4",
    "pillow>=12.0.0",
    "requests>=2.32.5",
    "rich>=14.2.0",
]
